Create hassio network layer / allow linking
